pyproject.toml
[project]
name = "parrot-olympe"
version = "0.0.0"  # the actual version is set by the release CI job
description = "Python controller library for Parrot Drones"
readme.file = "README.md"
readme.content-type = "text/markdown"
requires-python = ">=3.6"
license.text = "BSD 3-Clause Modification"
license.file = "LICENSE.md"
keywords = ["uav", "parrot", "drones", "parrot-drones"]
authors = [
  {name = "Nicolas Dessart"}
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "License :: OSI Approved",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: C",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Scientific/Engineering :: Image Processing",
  "Topic :: Scientific/Engineering"
]

dependencies = [
    "aenum",
    "asn1tools",
    "boltons",
    "colorlog",
    "dacite",
    "dataclasses; python_version == '3.6'",
    "future",
    "geopy",
    "ipython",
    "numpy",
    "pycryptodomex",
    "pytz",
    "PyYAML",
    "requests",
    "six",
    "tzlocal<4.0",
    "yapf",
    "websocket_client"
]

[project.optional-dependencies]
rendering = [
    "PyOpenGL",
    "PyOpenGL-accelerate",
    "PySDL2"
]

doc = [
    "Sphinx",
    "blockdiag",
    "furo",
    "seqdiag",
    "sphinx-copybutton",
    "sphinxcontrib-seqdiag"
]

tests = [
    "pytest"
]

[project.urls]
homepage = "developer.parrot.com/"
documentation = "developer.parrot.com/docs/olympe/"
repository = "github.com/Parrot-Developers/olympe"

[project.scripts]
olympe = "olympe.app:main"

[tool.alchemy]
top-levels = [
    "olympe",
    "olympe_deps",
    "arsdkparser.py",
    "arsdk",
    "google",
    "logness",
    "parrot",
    "ulog.py",
    "_ulog"
]

# Alchemy is not (yet) a python build system backend (See PEP 517 / PEP 518)
# [build-system]
# requires = ["alchemy-wheel>=1.3.8"]
# build-backend = "alchemy-wheel.api"
